+++ src/surface2.cpp
@@ -1,0 +1,4 @@
+#include "surface2.h"
+
+namespace batoid {
+}
